Corrected Command::argf() return type which was 'double' rather than 'float'. Added commands to change spotlight status, position, and colour. Added width and height as optional arguments to CA_SAVEBITMAP. Changed Canvas::configure() to accept integer width and height of new context - previously it would take this from the mainWidget_ but this resulted in the incorrect view being drawn for bitmaps of a different size to the main view. Added call to glColor3fv() before rotation globe is drawn, since it would not appear until the view was changed in some way.

+++ trunk/src/command/commandlist.cpp
@@ -188,7 +188,7 @@ double Command::argd(int argno)
 }
 
 // Return argument as float
-double Command::argf(int argno)
+float Command::argf(int argno)
 {
 	return (float) argd(argno);
 }
